% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{dr_pca}
\alias{dr_pca}
\title{PCA based data dimension reduction}
\usage{
dr_pca(dat, n_comp = NULL)
}
\arguments{
\item{dat}{[N, J] data matrix}

\item{n_comp}{[int] number of components to keep}

\item{control}{\{ ... \} list of runtime variables}
}
\value{
[N, n_comp] projected data
}
\description{
This is part of a data preprocessing pipeline for Paired Factor Analysis (pfa) where we
\itemize{
\item perform dimension reduction on data (determine the number of components to include via Broken Stick Model)
\item determine centers of major clusters to get initial factors
\item get initial weights of factor pairs by their relative distance, e.g. via projection on principle curve
}
}
\details{
\itemize{
\item n_comp If set to NULL, a broken stick model will be applied to determine the number of PCs to include
}
}
\examples{
...
}
\references{
...
}
\author{
Gao Wang and Kushal K. Dey
}
