% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.R
\name{get_model_lik}
\alias{get_model_lik}
\title{Get model likelihood for PFA}
\usage{
get_model_lik(D, F, Q, S)
}
\arguments{
\item{D}{[N, J] matrix of simulated data}

\item{F}{[K, J] true factor matrix}

\item{Q}{[N, 3] sample positions (columns are K1, K2, relative distance from K1)}

\item{S}{[J, 1] vector of simulated standard deviation for features}
}
\value{
loglik
}
\description{
This is a diagnostic function that computes likelihood given simulation parameters
}
\author{
Gao Wang and Kushal K. Dey
}
\references{
...
}

