% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{init_weight_princurve}
\alias{init_weight_princurve}
\title{Initialize weights for factor pairs by principle curve projection}
\usage{
init_weight_princurve(dat, factors)
}
\arguments{
\item{dat}{[N, J] data matrix, used to create principle curve}

\item{factors}{[K, J] factor matrix whose distance is to be calculated}

\item{control}{\{ ... \} list of runtime variables}
}
\value{
A list with elements below:
\item{factors}{...}
\item{weights}{...}
\item{lambda}{...}
\item{projected_points}{...}
}
\description{
This is part of a data preprocessing pipeline for Paired Factor Analysis (pfa) where we
\itemize{
\item perform dimension reduction on data (determine the number of components to include via Broken Stick Model)
\item determine centers of major clusters to get initial factors
\item get initial weights of factor pairs by their relative distance, e.g. via projection on principle curve
}
}
\details{
...
}
\examples{
...
}
\references{
...
}
\author{
Gao Wang and Kushal K. Dey
}
