% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc_transform.R
\name{pc_transform}
\alias{pc_transform}
\title{PCA based data transformation and factor initialization for pfa}
\usage{
pc_transform(dat, n_pc = NULL, control = NULL)
}
\arguments{
\item{dat}{[N, J] data matrix}

\item{n_pc}{[int] number of PC's to keep}

\item{control}{\{spins = 3\} list of runtime variables}
}
\value{
A list with elements below:
\item{data}{...}
\item{F}{...}
\item{P}{...}
}
\description{
This is a data preprocessing pipeline for Paired Factor Analysis (pfa) where we
\itemize{
\item perform PCA on data
\item determine the number of PCs to include via Broken Stick Model
\item determine centers of major clusters from the PC space via block model
\item project these centers to principle curve to produce initial values of factors and their weights for pfa
}
}
\details{
\itemize{
\item n_pc: if set to NULL, a broken stick model will be applied to determine the number of PCs to include
}
}
\examples{
...
}
\author{
Gao Wang and Kushal K. Dey
}
\references{
...
}

