% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.R
\name{pfa}
\alias{pfa}
\title{Paired factor analysis model}
\usage{
pfa(X, K = NULL, F = NULL, P = NULL, q = NULL, omega = NULL,
  control = NULL)
}
\arguments{
\item{X}{[N, J] matrix of observed data}

\item{K}{[int] number of factors}

\item{F}{[K, J] initial factor matrix}

\item{P}{[K, K] initial factor pair frequency matrix}

\item{q}{[C, 1] initial vector of possible membership loadings, a discrete set}

\item{omega}{[C, 1] initial weight of membership loadings, a discrete set corresponding to q}

\item{control}{\{tol = 1E-5, maxiter = 10000, logfile = NULL, n_cpu = 1\} list of runtime variables}
}
\value{
A list with elements below:
\item{F}{...}
\item{L}{...}
\item{P}{...}
\item{omega}{...}
\item{loglik}{...}
\item{loglik_diff}{...}
\item{niter}{...}
\item{status}{...}
}
\description{
...
}
\details{
...
}
\examples{
dat = readRDS('vignettes/example_data.rds')
dat = pfar::pc_transform(dat$X)
control = list(logfile = 'example_data.pfa', n_cpu = 8)
pfar::pfa(dat$data, F = dat$F, P = dat$P, control = control)
}
\author{
Gao Wang and Kushal K. Dey
}
\references{
...
}

