% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.R
\name{get_model_lik}
\alias{get_model_lik}
\title{Get model likelihood for PFA}
\usage{
get_model_lik(D, FCT, Qvec, Svec, pi_k, pi_q, log_scale = TRUE)
}
\arguments{
\item{D}{[N, J] matrix of simulated data}

\item{Qvec}{[C, 1] vector of simulated membership loadings, a discrete set}

\item{Svec}{[J, 1] vector of simulated standard deviation for features}

\item{pi_k}{[K * (K - 1) / 2, 1] vector of factor pair frequencies}

\item{pi_q}{[C, 1] vector of weights for simulated membership loadings}

\item{F}{[K, J] true factor matrix}
}
\description{
This is a diagnostic function that computes likelihood given simulation parameters
}
\author{
Gao Wang and Kushal K. Dey
}
\references{
...
}

