% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.R
\name{pfa}
\alias{pfa}
\title{Paired factor analysis model}
\usage{
pfa(X, K = NULL, F = NULL, P = NULL, q = NULL, alpha = NULL,
  control = NULL)
}
\arguments{
\item{X}{[N, J] matrix of observed data}

\item{K}{[int] number of factors}

\item{F}{[K, J] initial factor matrix}

\item{P}{[K, K] initial factor frequency matrix: the lower triangle is pair frequencies,
the diagonal is single factor frequencies.}

\item{q}{[C, 1] / [int] defines initial vector of possible membership loadings, a discrete set}

\item{alpha}{[double] prior for factor pair weights}

\item{control}{\{tol = 1E-5, maxiter = 10000, logfile = NULL, n_cpu = 4, verbose = FALSE \} list of runtime variables}
}
\value{
A list with elements below:
\item{F}{...}
\item{L}{...}
\item{P}{...}
\item{loglik}{...}
\item{loglik_diff}{...}
\item{niter}{...}
\item{status}{...}
}
\description{
...
}
\details{
\itemize{
\item q: When q is set to integer, q vector will be generated by seq(0, 100, by = max(as.integer(100 / q), 1)) / 100
}
}
\examples{
## The example data set can be installed via
## devtools::install_github("kkdey/singleCellRNASeqMouseDeng2014")
library(singleCellRNASeqMouseDeng2014)
meta_data <- pData(Deng2014MouseESC)
dat = exprs(Deng2014MouseESC)
K = 6
dat = t(limma::voom(dat)$E)
dr_dat = pfar::dr_pca(dat)
init_val = pfar::init_weight_princurve(dr_dat, pfar::init_factor_block(dr_dat, K)$factors)
control = list(logfile = 'example_data.pfa', n_cpu = 8)
res = pfar::pfa(dr_dat, F = init_val$factors, P = init_val$weights, control = control)
print(res)
}
\author{
Gao Wang and Kushal K. Dey
}
\references{
...
}

