\name{pfa}
\alias{pfa}
\title{Paired factor analysis model}
\description{...}

\usage{
pfa(X)
}

\arguments{
  \item{X}{[n,k] matrix of observed data}
}

\value{
  \item{loglik}{log likelihood after convergence}
}

\details{
  ...
}

\author{Gao Wang and Kushal K. Dey}

\references{
  ...
}

\examples{
library("pfar")
?pfa
}
