% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfa.R
\name{pfa}
\alias{pfa}
\title{Paired factor analysis model}
\usage{
pfa(X, K = NULL, F = NULL, P = NULL, q = NULL, alpha = NULL,
  control = NULL)
}
\arguments{
\item{X}{[N, J] matrix of observed data.}

\item{K}{[int] number of factors.}

\item{F}{[K, J] initial factor matrix.}

\item{P}{[K, K] initial factor frequency matrix: the lower triangle is pair frequencies,
the diagonal is single factor frequencies.}

\item{q}{[C, 1] / [int] defines initial vector of possible membership loadings, a discrete set, 0 < q < 1.}

\item{alpha}{[double] prior for factor pair weights.}

\item{control}{\{tol = 1E-4, maxiter = 10000, logfile = NULL, n_cpu = 4, init = c(20, 5), verbose = FALSE, init = NULL \} list of runtime variables}
}
\value{
A list with elements below:
\item{F}{...}
\item{L}{...}
\item{P}{...}
\item{alpha}{...}
\item{loglik}{...}
\item{BIC}{...}
\item{niter}{...}
\item{status}{...}
\item{runtime}{...}
}
\description{
...
}
\details{
\itemize{
\item q: When q is set to integer, q vector will be generated by seq(1, 100, by = max(as.integer(100 / q), 1)) / 100
}
}
\examples{
## see https://gaow.github.io/pfar
}
\references{
...
}
\author{
Gao Wang and Kushal K. Dey
}
